refact: move Actor-related structures to core.auth.actor
