+++ core-test/src/main/java/com/jeanbarrossilva/mastodonte/core/test/TestActorProvider.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.mastodonte.core.test
 
-import com.jeanbarrossilva.mastodonte.core.auth.Actor
-import com.jeanbarrossilva.mastodonte.core.auth.ActorProvider
+import com.jeanbarrossilva.mastodonte.core.auth.actor.Actor
+import com.jeanbarrossilva.mastodonte.core.auth.actor.ActorProvider
 
 /** [ActorProvider] that remembers and retrieves [Actor]s locally. **/
 class TestActorProvider : ActorProvider() {
