+++ core/sample/src/main/java/com/jeanbarrossilva/mastodonte/core/sample/auth/SampleAuthenticator.kt
@@ -1,8 +1,8 @@
 package com.jeanbarrossilva.mastodonte.core.sample.auth
 
-import com.jeanbarrossilva.mastodonte.core.auth.Actor
-import com.jeanbarrossilva.mastodonte.core.auth.ActorProvider
 import com.jeanbarrossilva.mastodonte.core.auth.Authenticator
+import com.jeanbarrossilva.mastodonte.core.auth.actor.Actor
+import com.jeanbarrossilva.mastodonte.core.auth.actor.ActorProvider
 
 /** [Authenticator] that provides a sample [Actor]. **/
 class SampleAuthenticator(override val actorProvider: ActorProvider) : Authenticator() {
