+++ core/shared-preferences/src/main/java/com/jeanbarrossilva/mastodonte/core/sharedpreferences/actor/SharedPreferencesActorProvider.kt
@@ -1,12 +1,12 @@
-package com.jeanbarrossilva.mastodonte.core.sharedpreferences
+package com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor
 
 import android.content.Context
 import android.content.SharedPreferences
 import androidx.core.content.edit
-import com.jeanbarrossilva.mastodonte.core.auth.Actor
-import com.jeanbarrossilva.mastodonte.core.auth.ActorProvider
-import com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor.MirroredActor
-import com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor.toMirroredActor
+import com.jeanbarrossilva.mastodonte.core.auth.actor.Actor
+import com.jeanbarrossilva.mastodonte.core.auth.actor.ActorProvider
+import com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor.mirror.MirroredActor
+import com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor.mirror.toMirroredActor
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
