+++ core/shared-preferences/src/main/java/com/jeanbarrossilva/mastodonte/core/sharedpreferences/actor/mirror/MirroredActor.kt
@@ -1,6 +1,6 @@
-package com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor
+package com.jeanbarrossilva.mastodonte.core.sharedpreferences.actor.mirror
 
-import com.jeanbarrossilva.mastodonte.core.auth.Actor
+import com.jeanbarrossilva.mastodonte.core.auth.actor.Actor
 import kotlinx.serialization.Serializable
 
 @Serializable
